add KnotController
