+++ Assets/Materials/Board/Trunk.prefab
@@ -50,8 +50,8 @@ MonoBehaviour:
       serializedVersion: 2
       m_Curve:
       - serializedVersion: 3
-        time: -0.049494803
-        value: 0.26729703
+        time: -0.020555854
+        value: 0.28211182
         inSlope: 0.024885133
         outSlope: 0.024885133
         tangentMode: 34
@@ -59,8 +59,8 @@ MonoBehaviour:
         inWeight: 0
         outWeight: 0.33333334
       - serializedVersion: 3
-        time: 1.0005065
-        value: 0.29342645
+        time: 1.0294455
+        value: 0.30824125
         inSlope: 0.024885133
         outSlope: 0.024885133
         tangentMode: 34
@@ -71,141 +71,6 @@ MonoBehaviour:
       m_PostInfinity: 2
       m_RotationOrder: 4
     m_Orientation: 152.8
-    m_TimeOfDeath: 0.1
-    m_Radius: 0.9
-  - m_HeightMap:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 3
-        time: -0.04306394
-        value: 0.32840815
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0
-        outWeight: 0.33333334
-      - serializedVersion: 3
-        time: 1.0069374
-        value: 0.35453758
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0.33333334
-        outWeight: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    m_Orientation: 143.3
-    m_TimeOfDeath: 0.1
-    m_Radius: 0.9
-  - m_HeightMap:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 3
-        time: -0.04627937
-        value: 0.24507481
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0
-        outWeight: 0.33333334
-      - serializedVersion: 3
-        time: 1.003722
-        value: 0.27120423
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0.33333334
-        outWeight: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    m_Orientation: 138.4
-    m_TimeOfDeath: 0.1
-    m_Radius: 0.9
-  - m_HeightMap:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 3
-        time: -0.03020221
-        value: 0.21544518
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0
-        outWeight: 0.33333334
-      - serializedVersion: 3
-        time: 1.0197991
-        value: 0.2415746
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0.33333334
-        outWeight: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    m_Orientation: 165
-    m_TimeOfDeath: 0.1
-    m_Radius: 0.9
-  - m_HeightMap:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 3
-        time: -0.010909617
-        value: 0.17655629
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0
-        outWeight: 0.33333334
-      - serializedVersion: 3
-        time: 1.0390917
-        value: 0.20268571
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0.33333334
-        outWeight: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    m_Orientation: 191.1
-    m_TimeOfDeath: 0.1
-    m_Radius: 0.9
-  - m_HeightMap:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 3
-        time: -0.0012633204
-        value: 0.12470445
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0
-        outWeight: 0.33333334
-      - serializedVersion: 3
-        time: 1.048738
-        value: 0.15083387
-        inSlope: 0.024885133
-        outSlope: 0.024885133
-        tangentMode: 34
-        weightedMode: 0
-        inWeight: 0.33333334
-        outWeight: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    m_Orientation: 191.1
     m_TimeOfDeath: 0.1
     m_Radius: 0.9
   m_MaxHeight: 15
