+++ Assets/Materials/Board/board.mat
@@ -98,7 +98,7 @@ Material:
     - _Glossiness: 0.24
     - _GlossyReflections: 1
     - _KnotCount: 1
-    - _MaxHeight: 15
+    - _MaxHeight: 100
     - _MaxRadius: 2
     - _Metallic: 0
     - _MinRadius: 1
